Made the initial setup functional
The number of players can be selected. Players can now enter their name, pick their character, players are then added to the game. The GUI is not functional past this point. Part of Game settingUp() is now broken and a lot of Game will likely need to be revised to implement the Swing UI